requirements.txt
openai~=0.27.4
customtkinter~=5.1.2
keyboard~=0.13.5
Pillow~=9.4.0
pystray~=0.19.4